requirements.txt
brotlipy==0.7.0
cffi==1.11.5
dj-database-url==0.5.0
Django==1.10.3
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
Markdown==2.6.11
Pillow==5.0.0
psycopg2==2.7.4
pycparser==2.18
pytz==2018.3
virtualenv==15.1.0
whitenoise==3.3.1
